Move global utils to dedicated file in UITestsFoundation
